findScranMarkers_one_vs_all

Identify marker feats for all clusters in a one vs all manner based on scran's implementation of findMarkers.

findScranMarkers_one_vs_all(
  gobject,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  subset_clusters = NULL,
  pval = 0.01,
  logFC = 0.5,
  min_feats = 10,
  min_genes = NULL,
  verbose = TRUE,
  ...
)

Arguments

gobject
giotto object

feat_type
feature type

expression_values
feat expression values to use

cluster_column
clusters to use

subset_clusters
subset of clusters to use

pval
filter on minimal p-value

logFC
filter on logFC

min_feats
minimum feats to keep per cluster, overrides pval and logFC

min_genes
deprecated, use min_feats

verbose
be verbose

...
additional parameters for the findMarkers function in scran

Value

data.table with marker feats

See also

Examples

data(mini_giotto_single_cell)

scran_markers = findScranMarkers_one_vs_all(gobject = mini_giotto_single_cell,
                                            cluster_column = 'leiden_clus')

#> using 'Scran' to detect marker feats. If used in published research, please cite:
#>   Lun ATL, McCarthy DJ, Marioni JC (2016).
#>   'A step-by-step workflow for low-level analysis of single-cell RNA-seq data with Bioconductor.'
#>   F1000Res., 5, 2122. doi: 10.12688/f1000research.9501.2. 

#> Error in findScranMarkers_one_vs_all(gobject = mini_giotto_single_cell,     cluster_column = "leiden_clus"): no slot of name "expression_feat" for this object of class "giotto"