mergeClusters

Merge selected clusters based on pairwise correlation scores and size of cluster.

mergeClusters(
  gobject,
  feat_type = NULL,
  expression_values = c("normalized", "scaled", "custom"),
  cluster_column,
  cor = c("pearson", "spearman"),
  new_cluster_name = "merged_cluster",
  min_cor_score = 0.8,
  max_group_size = 20,
  force_min_group_size = 10,
  max_sim_clusters = 10,
  return_gobject = TRUE,
  verbose = TRUE
)

Arguments

gobject
giotto object

feat_type
feature type

expression_values
expression values to use

cluster_column
name of column to use for clusters

cor
correlation score to calculate distance

new_cluster_name
new name for merged clusters

min_cor_score
min correlation score to merge pairwise clusters

max_group_size
max cluster size that can be merged

force_min_group_size
size of clusters that will be merged with their most similar neighbor(s)

max_sim_clusters
maximum number of clusters to potentially merge to reach force_min_group_size

return_gobject
return giotto object

verbose
be verbose

Value

Giotto object

Details

Merge selected clusters based on pairwise correlation scores and size of cluster. To avoid large clusters to merge the max_group_size can be lowered. Small clusters can be forcibly merged with their most similar pairwise cluster by adjusting the force_min_group_size parameter. Clusters smaller than this value will be merged independent on the provided min_cor_score value. The force_min_group_size might not always be reached if clusters have already been merged before
A giotto object is returned by default, if FALSE then the merging vector will be returned.

Examples

data("mini_giotto_single_cell")

pDataDT(mini_giotto_single_cell)

#> Error in pDataDT(mini_giotto_single_cell): no slot of name "expression_feat" for this object of class "giotto"

mini_giotto_single_cell = mergeClusters(mini_giotto_single_cell,
                                        cluster_column = 'leiden_clus',
                                        min_cor_score = 0.7,
                                        force_min_group_size = 4)

#> Error in mergeClusters(mini_giotto_single_cell, cluster_column = "leiden_clus",     min_cor_score = 0.7, force_min_group_size = 4): no slot of name "expression_feat" for this object of class "giotto"

pDataDT(mini_giotto_single_cell)

#> Error in pDataDT(mini_giotto_single_cell): no slot of name "expression_feat" for this object of class "giotto"

plotUMAP_2D(mini_giotto_single_cell, cell_color = 'merged_cluster', point_size = 3)

#> Error in dimPlot2D_single(gobject = gobject, feat_type = feat_type, dim_reduction_to_use = dim_reduction_to_use,     dim_reduction_name = dim_reduction_name, dim1_to_use = dim1_to_use,     dim2_to_use = dim2_to_use, spat_enr_names = spat_enr_names,     show_NN_network = show_NN_network, nn_network_to_use = nn_network_to_use,     network_name = network_name, cell_color = cell_color, color_as_factor = color_as_factor,     cell_color_code = cell_color_code, cell_color_gradient = cell_color_gradient,     gradient_midpoint = gradient_midpoint, gradient_limits = gradient_limits,     select_cell_groups = select_cell_groups, select_cells = select_cells,     show_other_cells = show_other_cells, other_cell_color = other_cell_color,     other_point_size = other_point_size, show_cluster_center = show_cluster_center,     show_center_label = show_center_label, center_point_size = center_point_size,     center_point_border_col = center_point_border_col, center_point_border_stroke = center_point_border_stroke,     label_size = label_size, label_fontface = label_fontface,     edge_alpha = edge_alpha, point_shape = point_shape, point_size = point_size,     point_alpha = point_alpha, point_border_col = point_border_col,     point_border_stroke = point_border_stroke, title = title,     show_legend = show_legend, legend_text = legend_text, legend_symbol_size = legend_symbol_size,     background_color = background_color, axis_text = axis_text,     axis_title = axis_title, show_plot = show_plot, return_plot = return_plot,     save_plot = save_plot, save_param = save_param, default_save_name = default_save_name): no slot of name "expression_feat" for this object of class "giotto"